Added IWidgetSet and dbwidget plus removal of various warnings.
git-svn-id: file:///home/pi/svn/RS7@41 04a3551b-0df6-43f4-9fb6-bc03da9130d0

+++ dev_midirx/midirx.cpp
@@ -52,7 +52,7 @@ void DeviceMidiRX::Clock()
 	
 		if( amount_read > 0 )
 		{
-			printf("MidiRX %d: Got 0x%02x\n", instance_name, buf );
+			printf("MidiRX %d: Got 0x%02x\n", instance_name, (int)buf );
 	
 			bool was_data;
 			was_data = midi_reader->Input((int)buf );
